% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_directory.R
\name{.get_tts}
\alias{.get_tts}
\title{get timestamp for file when asked for}
\usage{
.get_tts(file)
}
\arguments{
\item{file}{file name}
}
\value{
the content of the response
}
\description{
get timestamp for file when asked for
}
