% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_directory.R
\name{.hash_dir}
\alias{.hash_dir}
\title{Create hash of all files in the input directory and stores result in a file
named \code{file.sha25} in the \code{output} directory}
\usage{
.hash_dir(input, output, recursive = FALSE)
}
\arguments{
\item{input}{directory to hash}

\item{output}{where hash file should be stored}

\item{recursive}{if \code{TRUE}, directory will be recursively scanned for files to hash}
}
\value{
invisibly the file name and path of the hash file \code{file.sha.256}
}
\description{
Create hash of all files in the input directory and stores result in a file
named \code{file.sha25} in the \code{output} directory
}
