% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_directory.R
\name{.hash_dir_tree}
\alias{.hash_dir_tree}
\title{hash a directory tree}
\usage{
.hash_dir_tree(root, output)
}
\arguments{
\item{root}{root of the directory tree to be hashed}

\item{output}{directory containing the hash files in the tree structure}
}
\value{
invisibly the file name and path of the hash file \code{tree.sha.256}
}
\description{
hash a directory tree
}
