% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_directory.R
\name{.hash_recursively}
\alias{.hash_recursively}
\title{calculate sha256 hashes recursively for all files in directory}
\usage{
.hash_recursively(root, hashfile = tempfile(pattern = "file.sha256."))
}
\arguments{
\item{root}{root of the directory tree to be hashed}

\item{output}{directory containing the hash files in the tree structure}
}
\value{
file name and path of the hash file \code{file.sha.256}
}
\description{
calculate sha256 hashes recursively for all files in directory
}
