% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_directory.R
\name{hash_directory}
\alias{hash_directory}
\title{Create hash file of input directory and save to file in output directory}
\usage{
hash_directory(input, output)
}
\arguments{
\item{input}{dircetory of which the hashes should be calculated}

\item{output}{directory in which the hash files should be written}
}
\value{
invisibly \code{TRUE}
}
\description{
This function is calculating the \bold{file hash} for each subdirectory and
file and storing it in a \code{files.sha256} file and finally calculates the
\bold{directory hash} for this file and saves it in the file
\code{dir.sh256}. It is calculating one checksum file for the whole dataset
as well which is stored in the ToBeImported directory.
}
\examples{
\dontrun{
hash_directory(
  input = "./input",
  output = "./output"
)
}
}
