% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register the processing of o2meter data in the LEEF package}
\usage{
register(compression)
}
\arguments{
\item{compression}{the compression algorithm to be used. This can be:
\itemize{
\item \emph{none}: no compression, multiple files (copy of directory to be archived) - the fastest
\item \emph{tar}: no compression, single file - intermediate speed
\item \emph{tar.gz.max}: gz compression level 9 - very slow
}}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Register the processing of o2meter data in the LEEF package
}
