% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archive.none.R, R/run_archive.tar.R,
%   R/run_archive.tar.gz.R
\name{run_archive}
\alias{run_archive}
\title{Archive all files and folders in \code{input} to \code{output} without compression}
\usage{
run_archive(input, output)

run_archive(input, output)

run_archive(input, output)
}
\arguments{
\item{input}{directory which to archive}

\item{output}{directory in which the arcival directory will be created}

\item{compression}{which compression should be used for the archive. Thge following values are supported at the moment:
\describe{
   \item{\code{"none"}}{copy the to be imported data folder into archive}
   \item{\code{"tar"}}{tar the to be imported data folder}
   \item{\code{"tar.gz"}}{tar and gz compress the to be imported data folder}
}
Default is the value as defined in the config file.}

\item{get_tts}{if \code{TRUE}, a Trusted Time Stamp will be obtained from OriginStamp}
}
\value{
invisibly returns the name of the archivedir

invisibly returns the name of the archivefile

invisibly returns the name of the archivefile
}
\description{
Compress all files in \code{DATA_options("new_data_path")} into the directory \code{"NEW_DATA_PATH/../archive"}

Compress all files in \code{DATA_options("new_data_path")} into the directory \code{"NEW_DATA_PATH/../archive"}
}
\examples{
\dontrun{
run_archive.none(
  input = "./input",
  output = "./output"
)
}
\dontrun{
run_archive(
  compression = "tar",
  get_tts = TRUE
)
}
\dontrun{
run_archive(
  compression = "tar",
  get_tts = TRUE
)
}
}
