% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archive_none.R
\name{run_archive_none}
\alias{run_archive_none}
\title{Archive all files and folders in \code{input} to \code{output} without compression}
\usage{
run_archive_none(input, output)
}
\arguments{
\item{input}{directory which to archive}

\item{output}{directory in which the arcival directory will be created}
}
\value{
invisibly returns the name of the archivedir
}
\description{
Archive all files and folders in \code{input} to \code{output} without compression
}
\examples{
\dontrun{
run_archive_none(
  input = "./input",
  output = "./output"
)
}
}
