% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archive_tar.R
\name{run_archive_tar}
\alias{run_archive_tar}
\title{Compress and checksum the \code{DATA_options("new_data_path")}}
\usage{
run_archive_tar(input, output)
}
\arguments{
\item{compression}{which compression should be used for the archive. Thge following values are supported at the moment:
\describe{
   \item{\code{"none"}}{copy the to be imported data folder into archive}
   \item{\code{"tar"}}{tar the to be imported data folder}
   \item{\code{"tar.gz"}}{tar and gz compress the to be imported data folder}
}
Default is the value as defined in the config file.}

\item{get_tts}{if \code{TRUE}, a Trusted Time Stamp will be obtained from OriginStamp}
}
\value{
invisibly returns the name of the archivefile
}
\description{
Compress all files in \code{DATA_options("new_data_path")} into the directory \code{"NEW_DATA_PATH/../archive"}
}
\examples{
\dontrun{
run_archive(
  compression = "tar",
  get_tts = TRUE
)
}
}
