% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archive_none.R, R/run_archive_tar.R,
%   R/run_archive_tar.gz.R, R/run_archive_tar.gz_subdir.R,
%   R/run_archive_tar_subdir.R
\name{run_archive_none}
\alias{run_archive_none}
\alias{run_archive_tar}
\alias{run_archive_tar.gz}
\alias{run_archive_tar.gz_subdir}
\alias{run_archive_tar_subdir}
\title{Archive all files and folders in \code{input} to \code{output}}
\usage{
run_archive_none(input, output)

run_archive_tar(input, output)

run_archive_tar.gz(input, output)

run_archive_tar.gz_subdir(input, output)

run_archive_tar_subdir(input, output)
}
\arguments{
\item{input}{directory which to archive}

\item{output}{directory in which the archive will be created}
}
\value{
\code{run_archive_none()}: invisibly returns the name of the archive directory

\code{run_archive_tar()}: invisibly returns the name of the archive file

\code{run_archive_tar.gz()}: invisibly returns the name of the archive file

\code{run_archive_tar()}: invisibly returns the name of the archive file

\code{run_archive_tar()}: invisibly returns the name of the archive file
}
\description{
These functions archive the \code{input} folder, create checksums, and, depending on the function, compress the archive.
}
\examples{
\dontrun{
run_archive_none(
  input = "./input",
  output = "./output"
)
}
\dontrun{
run_archive.tar(
  input = "./input",
  output = "./output"
)
}
\dontrun{
run_archive.tar.gz(
  input = "./input",
  output = "./output"
)
}
\dontrun{
run_archive.tar(
  input = "./input",
  output = "./output"
)
}
\dontrun{
run_archive.tar(
  input = "./input",
  output = "./output"
)
}
}
